requirements.txt
opencv-python~=4.11.0.86
pyyaml~=6.0.2
pyqt5~=5.15.11
onnxruntime~=1.20.1
numpy~=2.1.1